✨ Sync padding with macOS window manager
