+++ Loop/Extensions/Defaults+Extensions.swift
@@ -89,6 +89,7 @@ extension Defaults.Keys {
     )
 
     // Advanced
+    static let useSystemWindowManagerWhenAvailable = Key<Bool>("useSystemWindowManagerWhenAvailable", default: true, iCloud: true)
     static let animateWindowResizes = Key<Bool>("animateWindowResizes", default: false, iCloud: true) // BETA
     static let disableCursorInteraction = Key<Bool>("disableCursorInteraction", default: false, iCloud: true)
     static let ignoreFullscreen = Key<Bool>("ignoreFullscreen", default: false, iCloud: true)
